+++ doc-examples/src/main/scala/elastic/FetchUsingSlickAndStreamIntoElastic.scala
@@ -4,22 +4,20 @@
 
 package elastic
 
-// format: off
 // #sample
-  import akka.Done
-  import akka.stream.alpakka.elasticsearch.{WriteIndexMessage, WriteMessage}
-  import akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSink
-  import org.apache.http.HttpHost
-  import org.elasticsearch.client.RestClient
-  import akka.stream.alpakka.slick.javadsl.SlickSession
-  import akka.stream.alpakka.slick.scaladsl.Slick
-  import spray.json.DefaultJsonProtocol._
-  import spray.json.JsonFormat
+import akka.Done
+import akka.stream.alpakka.elasticsearch.WriteMessage._
+import akka.stream.alpakka.elasticsearch.scaladsl.ElasticsearchSink
+import org.apache.http.HttpHost
+import org.elasticsearch.client.RestClient
+import akka.stream.alpakka.slick.javadsl.SlickSession
+import akka.stream.alpakka.slick.scaladsl.Slick
+import spray.json.DefaultJsonProtocol._
+import spray.json.JsonFormat
 
-  import scala.concurrent.Future
-  import scala.concurrent.duration._
+import scala.concurrent.Future
+import scala.concurrent.duration._
 // #sample
-// format: off
 
 import playground.elastic.ElasticsearchMock
 import playground.{ActorSystemAvailable, ElasticSearchEmbedded}
@@ -62,7 +60,7 @@ object FetchUsingSlickAndStreamIntoElastic extends ActorSystemAvailable with App
       .map {                                                                            // (7)
         case (id, genre, title, gross) => Movie(id, genre, title, gross)
       }
-      .map(movie => WriteMessage.createIndexMessage(movie.id.toString, movie))                     // (8)
+      .map(movie => createIndexMessage(movie.id.toString, movie))                       // (8)
       .runWith(ElasticsearchSink.create[Movie]("movie", "_doc"))                        // (9)
 
   done.onComplete {
